Замечание

+++ src/main/kotlin/lesson14/lesson-14-task-1.kt
@@ -2,22 +2,22 @@ package lesson14
 
 
 open class LinearSheep(
-    val speed: Double = 10.toDouble(),
-    val loadCapacity: Double = 10.toDouble(),
+    val speed: Double = 10.0,
+    val loadCapacity: Double = 10.0,
     val passengersAmount: Int = 100,
 )
 
 
 class CargoSheep(
-    speed: Double = 5.toDouble(),
-    loadCapacity: Double = 20.toDouble(),
+    speed: Double = 5.0,
+    loadCapacity: Double = 20.0,
     passengersAmount: Int = 40,
     ) : LinearSheep(speed, loadCapacity, passengersAmount)
 
 
 class Icebreaker(
-    speed: Double = 3.toDouble(),
-    loadCapacity: Double = 7.toDouble(),
+    speed: Double = 3.0,
+    loadCapacity: Double = 7.0,
     passengersAmount: Int = 20,
 ) : LinearSheep(speed, loadCapacity, passengersAmount) {
     val canBreakIce = true
